Dqn branch

+++ space_game/domain_names.py
@@ -4,11 +4,11 @@ from enum import Enum, auto
 
 KeyId = int
 Constraint = namedtuple('Constraint', ['min', 'max'])
-Coordinate = float
+Coordinate = int
 Coordinates = Tuple[Coordinate, Coordinate]
 Size = int
 Shape = Tuple[Size, Size]
-Acceleration = float
+Acceleration = int
 HitPoint = int
 ObjectId = int
 PlayerId = ObjectId
